% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dictionary.R
\docType{data}
\name{clean_dictionary}
\alias{clean_dictionary}
\title{Clean data dictionary.}
\format{
A data frame with 26 rows and 5 variables:
\describe{
\item{question_id}{The variable name}
\item{label}{Variable label}
\item{option_labels}{Labels of the factors}
\item{option_values}{Levels of the factors}
\item{type}{Variable type}
...
}
}
\usage{
clean_dictionary
}
\description{
A dataset on sample of clean data dictionary.
}
\keyword{datasets}
