% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{create_factors}
\alias{create_factors}
\title{Create factors from dictionary}
\usage{
create_factors(data, dictionary)
}
\arguments{
\item{data}{The dataset to create the factors}

\item{dictionary}{The data dictionary to pull value labels and levels}
}
\value{
tibble
}
\description{
Creates a factor variables from the data dictionary
}
\examples{
create_factors(datalabels::sample_data, datalabels::clean_dictionary)
}
