% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labels.R
\name{create_labels}
\alias{create_labels}
\title{Create labels of variables}
\usage{
create_labels(data, dictionary)
}
\arguments{
\item{data}{The dataset to create the factors}

\item{dictionary}{The data dictionary to pull variable labels}
}
\value{
tibble
}
\description{
Labels the variables from the data dictionary
}
\examples{
create_labels(datalabels::sample_data, datalabels::clean_dictionary)
}
