% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_clean.R
\name{dict_clean}
\alias{dict_clean}
\title{Clean the data dictionary}
\usage{
dict_clean(data)
}
\arguments{
\item{data}{Data dictionary to clean to clean}
}
\value{
tibble
}
\description{
Cleans the raw data dictionary
}
\examples{
dict_clean(data = datalabels::raw_dictionary)
}
