% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_dictionary.R
\docType{data}
\name{raw_dictionary}
\alias{raw_dictionary}
\title{Raw data dictionary.}
\format{
A data frame with 26 rows and 5 variables:
\describe{
\item{question_id}{The variable name}
\item{label}{Variable label}
\item{option_labels}{Labels of the factors}
\item{option_values}{Levels of the factors}
\item{type}{Variable type}
...
}
}
\usage{
raw_dictionary
}
\description{
A dataset on sample of raw data dictionary.
}
\keyword{datasets}
