% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Raw data dictionary.}
\format{
A data frame with 8 rows and 63 variables:
\describe{
\item{a1}{Gender}
\item{a2}{Age}
\item{a3}{Education level}
\item{a4}{Marital status}
\item{b1}{Employment status}
\item{b2}{Type of employment}
\item{b3}{Employment satisfaction}
\item{b4}{Having a business}
\item{d1}{Ever programmed}
\item{d2}{Programming language used}
...
}
}
\usage{
sample_data
}
\description{
A dataset on sample of raw data dictionary.
}
\keyword{datasets}
